Fix issue in de-duplication of assignments
In a (very unlikely) case an assignment generated for an EvalStmt
could end up in a sibling block of a block from which the statement
was being omitted. Address this by checking whether the assignment
occurs in a block dominating the block from which the statement is
being omitted, and if this is not the case, make a copy of the
expression being assigned and generate an EvalStmt for that copy.